ey-catalyst/gap-assessment/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.9.0
pydantic-settings==2.1.0
PyYAML==6.0.1
pandas==2.2.2
openpyxl==3.1.2
python-docx==1.0.1
requests==2.32.4
openai==1.51.0
PyPDF2==3.0.1
aiofiles==23.2.1

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Additional utilities
email-validator==2.1.0
